% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrv.forsite.app-package.R
\docType{package}
\name{nrv.forsite.app-package}
\alias{nrv.forsite.app}
\alias{nrv.forsite.app-package}
\title{nrv.forsite.app: Shiny App For Exploring FORSITE NRV Simulation Outputs}
\description{
Build and deploy FORSITE NRV shiny app.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FOR-CAST/nrv-forsite-app}
  \item Report bugs at \url{https://github.com/FOR-CAST/nrv-forsite-app/issues}
}

}
\author{
\strong{Maintainer}: Alex M. Chubaty \email{achubaty@for-cast.ca}

Authors:
\itemize{
  \item Peter Solymos \email{peter@analythium.io}
}

Other contributors:
\itemize{
  \item Michael Thomas \email{mthomas@ketchbrookanalytics.com} [contributor]
  \item Brad Lindblad \email{blindblad@ketchbrookanalytics.com} [contributor]
}

}
\keyword{internal}
