% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REGIONS}
\alias{REGIONS}
\title{Regions for NRV summary reporting}
\format{
named list of `sf` multipolygon objects
}
\usage{
REGIONS
}
\description{
a.k.a. reporting polygons
}
\keyword{datasets}
